Alberich integration: application controller refactoring, including removal of no-longer-needed ApplicationService file
